+++ rdlmpy/__init__.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-version_info = (0, 1, 'a3')
+version_info = (0, 1, 'a4')
 __version__ = ".".join([str(x) for x in version_info])
 
 from rdlmpy.client import RDLMClient
 from rdlmpy.context import RDLMContextManager
 from rdlmpy.lock import RDLMLock, RDLMActiveLock, RDLMWaitingLock
-from rdlmpy.exceptions import RDLMException, RDLMLockWaitExceededException, RDLMLockDeletedException, RDLMLockServerException
+from rdlmpy.exceptions import RDLMException, RDLMLockWaitExceededException, RDLMLockDeletedException, RDLMServerException, RDLMClientException
 
-__all__ = ['RDLMLock', 'RDLMActiveLock', 'RDLMWaitingLock', 'RDLMClient', 'RDLMContextManager', 'RDLMException', 'RDLMLockWaitExceededException', 'RDLMLockDeletedException', 'RDLMLockServerException']
+__all__ = ['RDLMLock', 'RDLMActiveLock', 'RDLMWaitingLock', 'RDLMClient', 'RDLMContextManager', 'RDLMException', 'RDLMLockWaitExceededException', 'RDLMLockDeletedException', 'RDLMServerException', 'RDLMClientException']